Updates Bing connection to use resource ID
Fixes an issue where the Bing connection required a full resource ID instead of just the connection name due to a Foundry bug.
This change updates the `.env` generation and Bicep templates to use the full resource ID for the Bing connection.
Also updates model capacity and dependencies in the dev setup.